Added CORE files update
Added an option to choose only the core Wordpress files. It scans a default list of Wordpress files from version 5.4 and adds them to the tar. You still have the option to tar all files.

Also cleaned up some white spaces.

Still need to clean up the variables and comments, but it is working great.